+++ cmd/terramate/cli/run.go
@@ -41,17 +41,17 @@ const (
 	// ErrRunCanceled represents the error when the execution was canceled.
 	ErrRunCanceled errors.Kind = "execution canceled"
 
-	// ErrRunCommandNotFound represents the error when the command cannot be found
-	// in the system.
-	ErrRunCommandNotFound errors.Kind = "command not found"
+	// ErrRunCommandNotExecuted represents the error when the command was not executed for whatever reason.
+	ErrRunCommandNotExecuted errors.Kind = "command not found"
 
 	cloudSyncPreviewCICDWarning = "--sync-preview is only supported in GitHub Actions workflows or Gitlab CICD pipelines"
 )
 
 // stackRun contains a list of tasks to be run per stack.
 type stackRun struct {
-	Stack *config.Stack
-	Tasks []stackRunTask
+	Stack         *config.Stack
+	Tasks         []stackRunTask
+	SyncTaskIndex int // index of the task with sync options
 }
 
 // stackCloudRun is a stackRun, but with a single task, because the cloud API only supports
@@ -211,7 +211,8 @@ func (c *cli) runOnStacks() {
 	var err error
 	for _, st := range stacks {
 		run := stackRun{
-			Stack: st.Stack,
+			SyncTaskIndex: -1,
+			Stack:         st.Stack,
 			Tasks: []stackRunTask{
 				{
 					Cmd:                  c.parsedArgs.Run.Command,
@@ -385,9 +386,22 @@ func (c *cli) runAll(
 	err = sched.Run(func(run stackRun) error {
 		errs := errors.L()
 
+		failedTaskIndex := -1
+
 	tasksLoop:
-		for _, task := range run.Tasks {
+		for taskIndex, task := range run.Tasks {
 			acquireResource()
+
+			// For cloud sync, we always assume that there's a single task per stack.
+			cloudRun := stackCloudRun{Stack: run.Stack, Task: task}
+
+			select {
+			case <-cancelCtx.Done():
+				c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCanceled))
+				releaseResource()
+				continue tasksLoop
+			default:
+			}
 
 			if !opts.Quiet && !opts.ScriptRun {
 				printer.Stderr.Println(printPrefix + " Entering stack in " + run.Stack.String())
@@ -411,48 +425,55 @@ func (c *cli) runAll(
 					input, err := config.EvalInput(evalctx, in)
 					if err != nil {
 						errs.Append(errors.E(err, "failed to evaluate input block"))
+						c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCommandNotExecuted, err))
 						releaseResource()
-						if continueOnError {
-							break tasksLoop
+						failedTaskIndex = taskIndex
+						if !continueOnError {
+							cancel()
 						}
-						cancel()
-						return errs.AsError()
+						break tasksLoop
 					}
 					otherStack, found, err := c.stackManager().StackByID(input.FromStackID)
 					if err != nil {
 						errs.Append(errors.E(err, "populating stack inputs from stack.id %s", input.FromStackID))
+						c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCommandNotExecuted, err))
 						releaseResource()
-						if continueOnError {
-							break tasksLoop
+						failedTaskIndex = taskIndex
+						if !continueOnError {
+							cancel()
 						}
-						cancel()
-						return errs.AsError()
+						break tasksLoop
 					}
 					if !found {
-						errs.Append(errors.E(
+						err := errors.E(
 							"Stack %s needs output from stack ID %q but it cannot be found",
 							run.Stack.Dir,
-							input.FromStackID))
+							input.FromStackID)
 
+						errs.Append(err)
+
+						c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCommandNotExecuted, err))
 						releaseResource()
-						if continueOnError {
-							break tasksLoop
+						failedTaskIndex = taskIndex
+						if !continueOnError {
+							cancel()
 						}
-						cancel()
-						return errs.AsError()
+						break tasksLoop
 					}
 
 					logger.Debug().Msgf("Stack depends on outputs from stack %s", otherStack.Dir)
 
 					backend, ok := cfg.SharingBackend(input.Backend)
 					if !ok {
-						errs.Append(errors.E("backend %s not found", input.Backend))
+						err := errors.E("backend %s not found", input.Backend)
+						errs.Append(err)
+						c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCommandNotExecuted, err))
 						releaseResource()
-						if continueOnError {
-							break tasksLoop
+						failedTaskIndex = taskIndex
+						if !continueOnError {
+							cancel()
 						}
-						cancel()
-						return errs.AsError()
+						break tasksLoop
 					}
 					_, ok = allOutputs[otherStack.Dir]
 					if !ok {
@@ -471,12 +492,13 @@ func (c *cli) runAll(
 						if err != nil {
 							if !task.MockOnFail {
 								errs.Append(errors.E(err, "failed to execute: %s (stderr: %s)", cmd.String(), stderr.Bytes()))
+								c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCommandNotExecuted, err))
 								releaseResource()
-								if continueOnError {
-									break tasksLoop
+								failedTaskIndex = taskIndex
+								if !continueOnError {
+									cancel()
 								}
-								cancel()
-								return errs.AsError()
+								break tasksLoop
 							}
 
 							printer.Stderr.Warnf("failed to execute `sharing_backend` command: %v", err)
@@ -484,65 +506,70 @@ func (c *cli) runAll(
 							stdoutBytes := stdout.Bytes()
 							typ, err := json.ImpliedType(stdoutBytes)
 							if err != nil {
-								errs.Append(errors.E(err, "unmashaling sharing_backend output"))
+								err := errors.E(err, "unmashaling sharing_backend output")
+								errs.Append(err)
+								c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCommandNotExecuted, err))
 								releaseResource()
-								if continueOnError {
-									break tasksLoop
+								failedTaskIndex = taskIndex
+								if !continueOnError {
+									cancel()
 								}
-								cancel()
-								return errs.AsError()
+								break tasksLoop
 							}
 							inputVal, err = json.Unmarshal(stdoutBytes, typ)
 							if err != nil {
-								errs.Append(errors.E(err, "unmashaling sharing_backend output"))
+								err := errors.E(err, "unmashaling sharing_backend output")
+								errs.Append(err)
+								c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCommandNotExecuted, err))
 								releaseResource()
-								if continueOnError {
-									break tasksLoop
+								failedTaskIndex = taskIndex
+								if !continueOnError {
+									cancel()
 								}
-								cancel()
-								return errs.AsError()
+								break tasksLoop
 							}
 						}
 						allOutputs[otherStack.Dir][backend.Name] = inputVal
 					}
 					stackOutputs := allOutputs[otherStack.Dir][backend.Name]
 					evalctx.SetNamespaceRaw("outputs", stackOutputs)
-					inputVal, err := input.Value(evalctx)
-					if err != nil {
-						if !task.MockOnFail || input.Mock.IsNull() {
-							errs.Append(errors.E(err, "evaluating input value"))
-							releaseResource()
-							if continueOnError {
-								break tasksLoop
+					inputVal, inputErr := input.Value(evalctx)
+					mockVal, mockFound, mockErr := input.Mock(evalctx)
+
+					if inputErr != nil {
+						if !task.MockOnFail || !mockFound || mockErr != nil {
+							err := errors.E(inputErr, "evaluating input value")
+							errs.Append(err)
+							if mockErr != nil {
+								errs.Append(errors.E(mockErr, "failed to evaluate input mock"))
 							}
-							cancel()
-							return errs.AsError()
-						}
-						inputVal = input.Mock
-					}
-					if !inputVal.Type().Equals(cty.String) {
-						errs.Append(errors.E("output type is not string but %s", inputVal.Type().FriendlyName()))
-						releaseResource()
-						if continueOnError {
+							c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCommandNotExecuted, err))
+							releaseResource()
+							failedTaskIndex = taskIndex
+							if !continueOnError {
+								cancel()
+							}
 							break tasksLoop
 						}
-						cancel()
-						return errs.AsError()
+
+						inputVal = mockVal
+					}
+					if !inputVal.Type().Equals(cty.String) {
+						err := errors.E("output type is not string but %s", inputVal.Type().FriendlyName())
+						errs.Append(err)
+						c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCommandNotExecuted, err))
+						releaseResource()
+						failedTaskIndex = taskIndex
+						if !continueOnError {
+							cancel()
+						}
+						break tasksLoop
 					}
 					environ = append(environ, stdfmt.Sprintf("TF_VAR_%s=%s", input.Name, inputVal.AsString()))
 				}
 			}
 
-			// For cloud sync, we always assume that there's a single task per stack.
-			cloudRun := stackCloudRun{Stack: run.Stack, Task: task, Env: environ}
-
-			select {
-			case <-cancelCtx.Done():
-				c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCanceled))
-				releaseResource()
-				continue
-			default:
-			}
+			cloudRun.Env = environ
 
 			cmdStr := strings.Join(task.Cmd, " ")
 			logger = logger.With().
@@ -551,14 +578,14 @@ func (c *cli) runAll(
 
 			cmdPath, err := runutil.LookPath(task.Cmd[0], environ)
 			if err != nil {
-				c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCommandNotFound, err))
+				c.cloudSyncAfter(cloudRun, runResult{ExitCode: -1}, errors.E(ErrRunCommandNotExecuted, err))
 				errs.Append(errors.E(err, "running `%s` in stack %s", cmdStr, run.Stack.Dir))
 				releaseResource()
-				if continueOnError {
-					break tasksLoop
+				failedTaskIndex = taskIndex
+				if !continueOnError {
+					cancel()
 				}
-				cancel()
-				return errs.AsError()
+				break tasksLoop
 			}
 
 			cmd := exec.Command(cmdPath, task.Cmd[1:]...)
@@ -610,11 +637,11 @@ func (c *cli) runAll(
 				errs.Append(errors.E(err, "running %s (at stack %s)", cmd, run.Stack.Dir))
 
 				releaseResource()
-				if continueOnError {
-					break tasksLoop
+				failedTaskIndex = taskIndex
+				if !continueOnError {
+					cancel()
 				}
-				cancel()
-				return errs.AsError()
+				break tasksLoop
 			}
 
 			resultc := makeResultChannel(cmd)
@@ -635,8 +662,13 @@ func (c *cli) runAll(
 					FinishedAt: &endTime,
 				}
 				c.cloudSyncAfter(cloudRun, res, errors.E(ErrRunCanceled))
+				errs.Append(errors.E(ErrRunCanceled, "execution aborted by CTRL-C (3x)"))
 				releaseResource()
-				return errors.E(ErrRunCanceled, "execution aborted by CTRL-C (3x)")
+				failedTaskIndex = taskIndex
+				if !continueOnError {
+					cancel()
+				}
+				break tasksLoop
 
 			case result := <-resultc:
 				logSyncWait()
@@ -664,17 +696,22 @@ func (c *cli) runAll(
 
 				c.cloudSyncAfter(cloudRun, res, err)
 				releaseResource()
-			}
-
-			err = errs.AsError()
-			if err != nil {
-				if continueOnError {
-					return err
+				if err != nil {
+					failedTaskIndex = taskIndex
+					if !continueOnError {
+						cancel()
+					}
+					break tasksLoop
 				}
-				logger.Info().Msg("interrupting execution of further stacks")
-				cancel()
-				return err
 			}
+		}
+
+		if failedTaskIndex != -1 && run.SyncTaskIndex != -1 && failedTaskIndex < run.SyncTaskIndex {
+			cloudRun := stackCloudRun{
+				Stack: run.Stack,
+				Task:  run.Tasks[run.SyncTaskIndex],
+			}
+			c.cloudSyncAfter(cloudRun, runResult{ExitCode: 1}, errors.E(ErrRunFailed))
 		}
 
 		return errs.AsError()